requirements.txt
argon2-cffi>=25.1.0
fastapi>=0.135.0
uvicorn>=0.44.0
httpx>=0.28.0
psutil>=7.0.0
python-dotenv>=1.2.0
itsdangerous>=2.2.0
python-multipart>=0.0.24
jinja2>=3.1.0
